+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Improved
 
-* Better APIs.
+* Better APIs and compile-time check for shared copies.
 
 ## [0.1.0] - 2025-02-04